Move ConditionEvaluator.cs to SIM.Pipelines
